requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==10.2.0
reportlab==4.1.0